[3.5.7.1] PageResponse total 属性类型修改

+++ cola-components/cola-component-dto/src/main/java/com/szmengran/cola/dto/PageResponse.java
@@ -109,7 +109,7 @@ public class PageResponse<T> extends Response {
         return response;
     }
 
-    public static <T> PageResponse<T> of(Collection<T> data, int totalCount, int pageSize, int pageIndex) {
+    public static <T> PageResponse<T> of(Collection<T> data, long totalCount, int pageSize, int pageIndex) {
         PageResponse<T> response = new PageResponse<>();
         response.setSuccess(true);
         response.setData(data);
